extract setup for install event loop
